requirements.txt
requests~=2.27.1
pdfplumber~=0.7.1
googleapis-common-protos~=1.56.4
google-pasta~=0.2.0
google-auth~=2.10.0
google-auth-oauthlib~=0.5.2
google-auth-httplib2~=0.1.0
google-api-python-client~=2.56.0
google-api-core~=2.8.2
PyYAML==6.0
PyQt5~=5.15.7